+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/like/LikeAPI.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.infrastructure.web.like;
 
-import com.mobble.mobbleserver.refactor.like.baseLike.dto.response.LikeMemberListResponseDto;
-import com.mobble.mobbleserver.refactor.like.baseLike.dto.response.LikeToggleResponseDto;
+import com.mobble.mobbleserver.infrastructure.web.like.dto.response.LikeMemberListResponseDto;
+import com.mobble.mobbleserver.infrastructure.web.like.dto.response.LikeToggleResponseDto;
 import com.mobble.mobbleserver.refactor.like.baseLike.entity.LikeType;
 import com.mobble.mobbleserver.refactor.like.baseLike.service.LikeDispatcherService;
 import lombok.RequiredArgsConstructor;
