+++ src/main/java/com/mobble/mobbleserver/refactor/like/articleLike/service/ArticleLikeQueryService.java
@@ -4,7 +4,7 @@ import com.mobble.mobbleserver.refactor.article.entity.Article;
 import com.mobble.mobbleserver.refactor.article.validator.ArticleValidator;
 import com.mobble.mobbleserver.refactor.like.articleLike.entity.ArticleLike;
 import com.mobble.mobbleserver.refactor.like.articleLike.repository.ArticleLikeRepository;
-import com.mobble.mobbleserver.refactor.like.baseLike.dto.response.LikeMemberListResponseDto;
+import com.mobble.mobbleserver.infrastructure.web.like.dto.response.LikeMemberListResponseDto;
 import com.mobble.mobbleserver.refactor.like.baseLike.entity.LikeType;
 import com.mobble.mobbleserver.refactor.like.baseLike.service.LikeQueryService;
 import lombok.RequiredArgsConstructor;
