+++ src/main/java/com/mobble/mobbleserver/refactor/like/baseLike/service/AbstractLikeService.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.refactor.like.baseLike.service;
 
-import com.mobble.mobbleserver.refactor.like.baseLike.dto.response.LikeToggleResponseDto;
+import com.mobble.mobbleserver.infrastructure.web.like.dto.response.LikeToggleResponseDto;
 import com.mobble.mobbleserver.refactor.like.baseLike.entity.LikeType;
 import com.mobble.mobbleserver.domain.member.Member;
 import lombok.RequiredArgsConstructor;
