+++ src/main/java/com/mobble/mobbleserver/refactor/like/baseLike/service/LikeDispatcherService.java
@@ -4,8 +4,8 @@ import com.mobble.mobbleserver.application.member.port.required.MemberReadPort;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.like.LikeErrorCode;
 import com.mobble.mobbleserver.global.exception.errorCode.member.MemberErrorCode;
-import com.mobble.mobbleserver.refactor.like.baseLike.dto.response.LikeMemberListResponseDto;
-import com.mobble.mobbleserver.refactor.like.baseLike.dto.response.LikeToggleResponseDto;
+import com.mobble.mobbleserver.infrastructure.web.like.dto.response.LikeMemberListResponseDto;
+import com.mobble.mobbleserver.infrastructure.web.like.dto.response.LikeToggleResponseDto;
 import com.mobble.mobbleserver.refactor.like.baseLike.entity.LikeType;
 import com.mobble.mobbleserver.domain.member.Member;
 import org.springframework.stereotype.Service;
