+++ src/main/java/com/mobble/mobbleserver/refactor/like/baseLike/service/LikeQueryService.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.refactor.like.baseLike.service;
 
-import com.mobble.mobbleserver.refactor.like.baseLike.dto.response.LikeMemberListResponseDto;
+import com.mobble.mobbleserver.infrastructure.web.like.dto.response.LikeMemberListResponseDto;
 import com.mobble.mobbleserver.refactor.like.baseLike.entity.LikeType;
 
 public interface LikeQueryService {
